requirements.txt
pytorch-lightning>=1.0.6
pytorch-lightning-bolts>=0.2.5
wandb>=0.10.10
PyYAML>=5.3.1
efficientnet-pytorch>=0.7.0
opencv-python>=4.4.0
torchvision~=0.8.1
pillow~=8.0.1
numpy~=1.19.2
streamlit~=0.71.0
altair~=4.1.0
pandas~=1.1.4
matplotlib~=3.3.2
decord~=0.4.2